+++ src/QuantitativeWorld/PowerUnit.cs
@@ -7,13 +7,13 @@ namespace QuantitativeWorld
     {
         private readonly string _name;
         private readonly string _abbreviation;
-        private readonly decimal? _valueInWatts;
+        private readonly double? _valueInWatts;
 
-        public PowerUnit(string name, string abbreviation, decimal valueInWatts)
+        public PowerUnit(string name, string abbreviation, double valueInWatts)
         {
             Assert.IsNotNullOrWhiteSpace(name, nameof(name));
             Assert.IsNotNullOrWhiteSpace(abbreviation, nameof(abbreviation));
-            Assert.IsGreaterThan(valueInWatts, 0m, nameof(valueInWatts));
+            Assert.IsGreaterThan(valueInWatts, 0d, nameof(valueInWatts));
 
             _name = name;
             _abbreviation = abbreviation;
@@ -22,10 +22,10 @@ namespace QuantitativeWorld
 
         public string Name => _name ?? Watt._name;
         public string Abbreviation => _abbreviation ?? Watt._abbreviation;
-        public decimal ValueInWatts => _valueInWatts ?? Watt._valueInWatts.Value;
+        public double ValueInWatts => _valueInWatts ?? Watt._valueInWatts.Value;
 
         public override string ToString() => Abbreviation;
 
-        decimal ILinearUnit.ValueInBaseUnit => ValueInWatts;
+        double ILinearUnit.ValueInBaseUnit => ValueInWatts;
     }
 }
